+++ geo/CHANGES.md
@@ -12,6 +12,8 @@
 * Fix: `ClosestPoint` for Polygon's handling of internal points
 * Implemented `ClosestPoint` method for types Triangle, Rect, GeometryCollection, Coordinate and the Geometry enum.
   * <https://github.com/georust/geo/pull/675>
+* BREAKING: `TryMapCoords` Result is now generic rather than a Box<dyn Error>.
+  * <https://github.com/georust/geo/issues/722>
 
 ## 0.18.0